+++ src/libs/naive_2dge/include/naive_2dge/image.hpp
@@ -35,7 +35,7 @@ namespace naive_2dge
 
         void* get_resource();
 
-        void get_size(std::uint32_t& w, std::uint32_t& h) const;
+        void get_size(std::uint16_t& w, std::uint16_t& h) const;
 
     private:
