+++ src/libs/naive_2dge/texture.cpp
@@ -70,7 +70,7 @@ namespace naive_2dge
 		return impl_->resource_;
 	}
 
-	void texture::get_size(std::uint32_t& w, std::uint32_t& h) const
+	void texture::get_size(std::uint16_t& w, std::uint16_t& h) const
 	{
 		w = impl_->w_;
 		h = impl_->h_;
